A_carrom_game.cpp
There is a table of carrom game which has 4 sides. There are 4 players sitting on each side in the following order, A,B,C and D.
The players sitting opposite to each other are in the same team. Therefore, A-C and B-D are the two teams.
There are N black pawns and N white pawns distributed in the table at different positions. And there is also one Red pawn. 
The game goes as follows, 
1. Each pawn of each color has its difficulty levels givens. Initially, A,B,C and D have their skill levels as P,Q,R and S respectively.
2. The player can only hit a certain pawn if and only if the difficulty level of that pawn is less than equal to the skill level of that player. After hitting that particular pawn, the skill level of that player increases by that difficulty level.

Now the task for the players is to hit the pawns. A-C team will always target Black Pawns and B-D will always target White Pawns. No one will target Red pawn. If all the N pawns assigned to that team are hit, then only they will pursue Red Pawn. Whoever successfully hits the red pawn wins the game.

For every turn, in which a player is unable to hit any pawn, his skill level would be incremented by one.

The turn of the players goes as follows:
A-B-C-D-A-B....and so on.
Here, A starts first.

Notes 
1. A&C are on the same team and so are B&D .
2. Red pawn will only be hit after every N pawns of assigned color to that team are destroyed.
3. A player can only hit a single pawn during his turn if he wishes to do so. He can also opt to skip his turn, that is, he can decide not to hit any pawn even if he is capable of hitting one.
4. If a player opts for a skip round, his skill level will still increase by 1 as he did not hit any pawn during his turn.
You have to determine who will win the game considering each team plays optimally.

SAMPLE INPUT 
1
2
1
0
1 1 1 1

SAMPLE OUTPUT 
B-D